backend/requirements.txt
# Django
django
djangorestframework
psycopg2-binary
django-cors-headers
